Some build fixes

+++ Sources/SolarTears/Platform/Win32/Win32Util.inl
@@ -1,10 +1,10 @@
 std::string ConvertWstringToUTF8(const std::wstring_view str)
 {
 	std::string result;
-	int charCount = WideCharToMultiByte(CP_UTF8, WC_COMPOSITECHECK, str.data(), str.size(), nullptr, 0, nullptr, nullptr);
+	int charCount = WideCharToMultiByte(CP_UTF8, WC_COMPOSITECHECK, str.data(), (int)str.size(), nullptr, 0, nullptr, nullptr);
 
 	result.resize(charCount);
-	WideCharToMultiByte(CP_UTF8, WC_COMPOSITECHECK, str.data(), str.size(), result.data(), (int)result.size(), nullptr, nullptr);
+	WideCharToMultiByte(CP_UTF8, WC_COMPOSITECHECK, str.data(), (int)str.size(), result.data(), (int)result.size(), nullptr, nullptr);
 
 	return result;
 }
@@ -12,10 +12,10 @@ std::string ConvertWstringToUTF8(const std::wstring_view str)
 std::wstring ConvertUTF8ToWstring(const std::string_view str)
 {
 	std::wstring result;
-	int charCount = MultiByteToWideChar(CP_UTF8, MB_COMPOSITE, str.data(), str.size(), nullptr, 0);
+	int charCount = MultiByteToWideChar(CP_UTF8, MB_COMPOSITE, str.data(), (int)str.size(), nullptr, 0);
 
 	result.resize(charCount);
-	MultiByteToWideChar(CP_UTF8, MB_COMPOSITE, str.data(), str.size(), result.data(), (int)result.size());
+	MultiByteToWideChar(CP_UTF8, MB_COMPOSITE, str.data(), (int)str.size(), result.data(), (int)result.size());
 
 	return result;
 }
